+++ features/roomdetails/impl/src/main/kotlin/io/element/android/features/roomdetails/impl/edit/RoomDetailsEditView.kt
@@ -42,7 +42,7 @@ import io.element.android.libraries.designsystem.theme.aliasScreenTitle
 import io.element.android.libraries.designsystem.theme.components.Scaffold
 import io.element.android.libraries.designsystem.theme.components.Text
 import io.element.android.libraries.designsystem.theme.components.TextButton
-import io.element.android.libraries.designsystem.theme.components.TextField2
+import io.element.android.libraries.designsystem.theme.components.TextField
 import io.element.android.libraries.designsystem.theme.components.TopAppBar
 import io.element.android.libraries.matrix.ui.components.AvatarActionBottomSheet
 import io.element.android.libraries.matrix.ui.components.EditableAvatarView
@@ -108,7 +108,7 @@ fun RoomDetailsEditView(
             )
             Spacer(modifier = Modifier.height(60.dp))
 
-            TextField2(
+            TextField(
                 label = stringResource(id = R.string.screen_room_details_room_name_label),
                 value = state.roomRawName,
                 placeholder = stringResource(CommonStrings.common_room_name_placeholder),
@@ -119,7 +119,7 @@ fun RoomDetailsEditView(
 
             Spacer(modifier = Modifier.height(28.dp))
 
-            TextField2(
+            TextField(
                 label = stringResource(CommonStrings.common_topic),
                 value = state.roomTopic,
                 placeholder = stringResource(CommonStrings.common_topic_placeholder),
